feat(TaskGrid): refactor localization handling and improve task management features
