+++ src/components/TaskGrid/components/header/view-switcher/view-manager/ViewManager.ts
@@ -1,7 +1,8 @@
 import { Dataset, ICommand, IDataProvider, IDataset, IRetrieveRecordCommandOptions } from "@talxis/client-libraries";
 import { DatasetControl, IDatasetControl } from "../../../../../../utils/dataset-control";
 import { ITaskGridDatasetControl } from "../../../../interfaces";
-import { ILocalizationService, ITaskGridLabels } from "../../../../labels";
+import { ILocalizationService } from "../../../../../../utils";
+import { ITaskGridLabels } from "../../../../labels";
 import { IDeletedUserQueriesResult, ISavedQueryDataProvider } from "../../../../data-providers/saved-query-data-provider";
 import { ErrorHelper } from "../../../../../../utils";
