+++ src/components/TaskGrid/interfaces.ts
@@ -4,8 +4,9 @@ import { IGridCustomizerStrategy } from "./components/grid/grid-customizer";
 import { ICustomColumnsDataProvider, ICustomColumnsStrategy } from "./data-providers/custom-columns-data-provider/CustomColumnsDataProvider";
 import { ISavedQueryDataProvider, ISavedQueryStrategy } from "./data-providers/saved-query-data-provider";
 import { ITaskDataProviderStrategy, ITaskDataProvider } from "./data-providers/task-data-provider";
-import { ILocalizationService, ITaskGridLabels } from "./labels";
+import { ITaskGridLabels } from "./labels";
 import { ITaskGridState } from "./TaskGridDatasetControlFactory";
+import { ILocalizationService } from "../../utils";
 
 export interface ITaskGridDatasetControlParameters {
     dataset: IDataset;
