+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './theme';
 export * from './BaseControls';
 export * from './dataset';
 export * from './error-handling';
+export * from './localization';
